rest

+++ reservation/src/main/java/ma/ensa/reservation/models/Chambre.java
@@ -11,7 +11,7 @@ public class Chambre {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Enumerated(EnumType.STRING)
     private TypeChambre typeChambre;
@@ -24,11 +24,11 @@ public class Chambre {
     public Chambre() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -57,6 +57,6 @@ public class Chambre {
     }
 
     public Chambre(Long chambreId) {
-        this.id = chambreId.intValue();
+        this.id = chambreId;
     }
 }
